requirements.txt
transformers==4.35.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.1.0
numpy==1.24.3